2019-12-20 React+Express(Node) SLP Token Upload Image
Needs to accept either a >=128x128px image or an SVG
The upload request should also include a SLP token id, and a signed message from the genesis address of that token id to prove ownership.
The signed message should be a sha256 hash of the submitted image file.
Obviously, the signature must be verified.
Upon accept, it should create a 32x32px version, a 64x64px version, and a 128x128px version and submit in a PR to https://github.com/kosinusbch/slp-token-icons. The original untouched upload should also be submitted along with those as, and the signed message should be included as a PR

Do not submit anything to that repo during any testing. Fork that repo and test in the fork

BCH libraries
https://bitcoincashjs.github.io/#Documentation